+++ wagtailsocialfeed/utils/feed/twitter.py
@@ -1,15 +1,20 @@
 import logging
-import json
 
 from dateutil import parser as dateparser
 from django.core.exceptions import ImproperlyConfigured
 from twython import Twython
-
 from wagtailsocialfeed.utils.conf import get_socialfeed_setting
 
-from . import AbstractFeed, FeedItem
+from . import AbstractFeed, AbstractFeedQuery, FeedItem
 
 logger = logging.getLogger('wagtailsocialfeed')
+
+settings = get_socialfeed_setting('CONFIG').get('twitter', None)
+if not settings:
+    raise ImproperlyConfigured(
+        "No twitter configuration defined in the settings. "
+        "Make sure you define WAGTAIL_SOCIALFEED_CONFIG in your "
+        "settings with at least a 'twitter' entry.")
 
 
 class TwitterFeedItem(FeedItem):
@@ -46,79 +51,54 @@ def process_images(media_dict):
     return images
 
 
+class TwitterFeedQuery(AbstractFeedQuery):
+    def __init__(self, username, query_string):
+        super(TwitterFeedQuery, self).__init__(username, query_string)
+
+        self.twitter = Twython(settings['CONSUMER_KEY'],
+                               settings['CONSUMER_SECRET'],
+                               settings['ACCESS_TOKEN_KEY'],
+                               settings['ACCESS_TOKEN_SECRET'])
+
+    def _search(self, tweet):
+        """Very basic search function"""
+        return self.query_string.lower() in tweet['text'].lower()
+
+    def __call__(self, max_id=None):
+        """
+        Return the raw data fetched from twitter and the oldest post in the
+        result set.
+
+        We return the oldest post as well because the raw data might be
+        filtered based on the query_string, but we still need the oldest
+        post to check the date to determine wether we should stop our
+        search or continue
+        """
+        raw = self.twitter.get_user_timeline(
+            screen_name=self.username,
+            trim_user=True,
+            contributor_details=False,
+            include_rts=False,
+            max_id=max_id)
+
+        if not raw:
+            return raw, None
+
+        oldest_post = raw[-1]
+        if self.query_string:
+            raw = list(filter(self._search, raw))
+        return raw, oldest_post
+
+
 class TwitterFeed(AbstractFeed):
-    def __init__(self, *args, **kwargs):
-        super(TwitterFeed, self).__init__(*args, **kwargs)
+    item_cls = TwitterFeedItem
+    query_cls = TwitterFeedQuery
 
-        self.settings = get_socialfeed_setting('CONFIG').get('twitter', None)
-        if not self.settings:
-            raise ImproperlyConfigured(
-                "No twitter configuration defined in the settings. "
-                "Make sure you define WAGTAIL_SOCIALFEED_CONFIG in your "
-                "settings with at least a 'twitter' entry.")
+    def get_post_date(self, item):
+        return dateparser.parse(item.get('created_at'))
 
-        self.twitter = Twython(self.settings['CONSUMER_KEY'],
-                               self.settings['CONSUMER_SECRET'],
-                               self.settings['ACCESS_TOKEN_KEY'],
-                               self.settings['ACCESS_TOKEN_SECRET'])
-
-    def fetch_online(self, config, limit=None, query_string=None, max_id=None):
-        username = config.username
-
-        def _search(tweet):
-            """Very basic search function"""
-            return query_string.lower() in tweet['text'].lower()
-
-        self.iteration = 0
-
-        def _fetch(max_id=None):
-            """
-            Return the raw data fetched from twitter and the oldest post in the
-            result set.
-
-            We return the oldest post as well because the raw data might be
-            filtered based on the query_string, but we still need the oldest
-            post to check the date to determine wether we should stop our
-            search or continue
-            """
-            raw = self.twitter.get_user_timeline(
-                screen_name=username,
-                trim_user=True,
-                contributor_details=False,
-                include_rts=False,
-                max_id=max_id)
-
-            if not raw:
-                return raw, None
-
-            oldest_post = raw[-1]
-            if query_string:
-                raw = list(filter(_search, raw))
-            return raw, oldest_post
-
-        raw, oldest_post = _fetch()
-
-        if query_string:
-            # If we have a query_string, we should fetch the users timeline a
-            # couple of times to dig a bit into the history.
-            while oldest_post:
-                # Trick from twitter API doc to exclude the oldest post from
-                # the next result-set
-                oldest_post_date = dateparser.parse(oldest_post.get(
-                    'created_at'))
-                if not self.more_history_allowed(oldest_post_date):
-                    break
-
-                max_id = oldest_post['id'] - 1
-                _raw, _post = _fetch(max_id=max_id)
-                if _post and _post['id'] == oldest_post['id']:
-                    logger.warning("Trying to fetch older tweets but received "
-                                   "same result set. Breaking the loop.")
-                    break
-                oldest_post = _post
-                raw += _raw
-
-        return raw
-
-    def to_feed_item(self, raw):
-        return TwitterFeedItem.from_raw(raw)
+    def fetch_older_items(self, query, oldest_post):
+        # Trick from twitter API doc to exclude the oldest post from
+        # the next result-set
+        max_id = oldest_post['id'] - 1
+        return query(max_id=max_id)